split backend_qt4 into two parts, with and without Gcf
added a static method to FigureManagerQt to gracefully patch in
the Gcf.destroy call back.  As with the _key_press_handler call back
this may be a stop-gap if that logic gets re-factored.
